fix(nextly): classify stored hook errors the same way
Auditing the examples turned up a third copy of the rebuild: the stored,
UI-configured executor also replaced every thrown error with a generic one,
so a stored hook rejecting input lost its status, code and field issues
exactly as the runtime path did.

The classification moves into its own module and both paths use it. A
rejection means the same thing whichever way the hook was declared, and a
second copy of this logic is free to drift from the first -- which is how
there came to be three.

The stored path passes the hook id as extra log context, so which stored
hook failed is still recorded without appearing in the caller's message.

Also migrates the remaining `beforeOperation` example, which still threw a
plain error for a rate limit, and adds the missing import to the plugin
example so it compiles when copied.
